requirements.txt
amqp==5.3.1
asgiref==3.10.0
bcrypt==4.1.1
billiard==4.2.2
celery==5.3.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
dj-database-url==3.0.1
Django==5.0.14
django-cors-headers==4.3.1
django-debug-toolbar==6.1.0
django-filter==23.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
dnspython==2.8.0
eventlet==0.40.3
greenlet==3.2.4
icalendar==5.0.10
idna==3.11
importlib-metadata==6.11.0
kombu==5.6.0
MarkupSafe==3.0.3
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.8.2
python-decouple==3.8
python-http-client==3.3.7
pytz==2025.2
redis==4.6.0
requests==2.32.5
sendgrid==6.12.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.11.0
zipp==3.23.0
